fix stats format complain. Add errno to build on MacOS

+++ src/stats.cpp
@@ -2,16 +2,16 @@
 
 void Stats::print_stats() {
     fprintf(stderr, "General Statistics\n");
-    fprintf(stderr, "  Packets processed: %10lu\n", packets_processed);
-    fprintf(stderr, "  Packets skipped:   %10lu (%.2f%%)\n", packets_skipped, 100.0 * (double(packets_skipped) / packets_processed));
-    fprintf(stderr, "  Packets parsed:    %10lu (%.2f%%)\n", packets_parsed, 100.0 * (double(packets_parsed) / packets_processed));
+    fprintf(stderr, "  Packets processed: %10llu\n", (unsigned long long) packets_processed);
+    fprintf(stderr, "  Packets skipped:   %10llu (%.2f%%)\n", (unsigned long long) packets_skipped, 100.0 * (double(packets_skipped) / packets_processed));
+    fprintf(stderr, "  Packets parsed:    %10llu (%.2f%%)\n", (unsigned long long) packets_parsed, 100.0 * (double(packets_parsed) / packets_processed));
     fprintf(stderr, "Network Layer Statistics (of packets parsed)\n");
-    fprintf(stderr, "  IPv4:              %10lu (%.2f%%)\n", ipv4, 100.0 * (double(ipv4) / packets_parsed));
-    fprintf(stderr, "  IPv6:              %10lu (%.2f%%)\n", ipv6, 100.0 * (double(ipv6) / packets_parsed));
+    fprintf(stderr, "  IPv4:              %10llu (%.2f%%)\n", (unsigned long long) ipv4, 100.0 * (double(ipv4) / packets_parsed));
+    fprintf(stderr, "  IPv6:              %10llu (%.2f%%)\n", (unsigned long long) ipv6, 100.0 * (double(ipv6) / packets_parsed));
     fprintf(stderr, "Transport Layer Statistics (of packets parsed)\n");
-    fprintf(stderr, "  TCP:               %10lu (%.2f%%)\n", tcp, 100.0 * (double(tcp) / packets_parsed));
-    fprintf(stderr, "  UDP:               %10lu (%.2f%%)\n", udp, 100.0 * (double(udp) / packets_parsed));
-    fprintf(stderr, "  ICMP:              %10lu (%.2f%%)\n", icmp, 100.0 * (double(icmp) / packets_parsed));
+    fprintf(stderr, "  TCP:               %10llu (%.2f%%)\n", (unsigned long long) tcp, 100.0 * (double(tcp) / packets_parsed));
+    fprintf(stderr, "  UDP:               %10llu (%.2f%%)\n", (unsigned long long) udp, 100.0 * (double(udp) / packets_parsed));
+    fprintf(stderr, "  ICMP:              %10llu (%.2f%%)\n", (unsigned long long) icmp, 100.0 * (double(icmp) / packets_parsed));
 }
 
 void Stats::update(bool parsed, uint8_t network_layer, uint8_t transport_layer) {
